+++ include/astarte_device_sdk/uuid.h
@@ -21,8 +21,14 @@
 /** @brief Number of bytes in the binary representation of a UUID. */
 #define ASTARTE_UUID_SIZE 16
 
-/** @brief Length of the UUID string representation. */
+/** @brief Length of the UUID canonical string representation. */
 #define ASTARTE_UUID_STR_LEN 36
+
+/** @brief Length of the UUID base64 string representation. */
+#define ASTARTE_UUID_BASE64_LEN 24
+
+/** @brief Length of the UUID base64 URL and filename safe string representation. */
+#define ASTARTE_UUID_BASE64URL_LEN 22
 
 /** @brief Binary representation of a UUID. */
 typedef uint8_t astarte_uuid_t[ASTARTE_UUID_SIZE];
@@ -50,11 +56,32 @@ astarte_result_t astarte_uuid_generate_v5(
  * @brief Convert a UUID to its canonical (RFC4122) string representation.
  *
  * @param[in] uuid The UUID to convert to string.
- * @param[out] out  A pointer to a previously allocated buffer where the result will be written.
+ * @param[out] out A pointer to a previously allocated buffer where the result will be written.
  * @param[in] out_size The size of the out buffer. Should be at least 37 bytes.
- * @return ASTARTE_RESULT_INTERNAL_ERROR upon failure, ASTARTE_RESULT_OK otherwise.
+ * @return ASTARTE_RESULT_OK on success, an error code otherwise.
  */
 astarte_result_t astarte_uuid_to_string(const astarte_uuid_t uuid, char *out, size_t out_size);
+
+/**
+ * @brief Convert a UUID to its base 64 (RFC 3548, RFC 4648) string representation.
+ *
+ * @param[in] uuid The UUID to convert to string.
+ * @param[out] out A pointer to a previously allocated buffer where the result will be written.
+ * @param[in] out_size The size of the out buffer. Should be at least 25 bytes.
+ * @return ASTARTE_RESULT_OK on success, an error code otherwise.
+ */
+astarte_result_t astarte_uuid_to_base64(const astarte_uuid_t uuid, char *out, size_t out_size);
+
+/**
+ * @brief Convert a UUID to its base 64 (RFC 4648 sec. 5) URL and filename safe string
+ * representation.
+ *
+ * @param[in] uuid The UUID to convert to string.
+ * @param[out] out A pointer to a previously allocated buffer where the result will be written.
+ * @param[in] out_size The size of the out buffer. Should be at least 23 bytes.
+ * @return ASTARTE_RESULT_OK on success, an error code otherwise.
+ */
+astarte_result_t astarte_uuid_to_base64url(const astarte_uuid_t uuid, char *out, size_t out_size);
 
 /**
  * @brief Parse a UUID from its canonical (RFC4122) string representation.