+++ lib/astarte_device_sdk/include/pairing_private.h
@@ -36,13 +36,14 @@ extern "C" {
  * @brief Fetch the MQTT broker URL from Astarte.
  *
  * @param[in] timeout_ms Timeout to use for the HTTP operations in ms.
+ * @param[in] device_id Unique identifier to use to register the device instance.
  * @param[in] cred_secr Credential secret to use as authorization token.
  * @param[out] out_url Output buffer where to store the fetched ULR.
  * @param[in] out_url_size Size of the output buffer for the URL.
  * @return ASTARTE_RESULT_OK if successful, otherwise an error code.
  */
-astarte_result_t astarte_pairing_get_broker_url(
-    int32_t timeout_ms, const char *cred_secr, char *out_url, size_t out_url_size);
+astarte_result_t astarte_pairing_get_broker_info(int32_t timeout_ms, const char *device_id,
+    const char *cred_secr, char *out_url, size_t out_url_size);
 
 /**
  * @brief Fetch the client x509 certificate from Astarte.
@@ -50,6 +51,7 @@ astarte_result_t astarte_pairing_get_broker_url(
  * @warning This is often a very memory intensive operation, more than 20kB of stack are required.
  *
  * @param[in] timeout_ms Timeout to use for the HTTP operations in ms.
+ * @param[in] device_id Unique identifier to use to register the device instance.
  * @param[in] cred_secr Credential secret to use as authorization token.
  * @param[out] privkey_pem Buffer where to store the computed private key, in the PEM format.
  * @param[in] privkey_pem_size Size of preallocated privkey_pem buffer.
@@ -57,8 +59,9 @@ astarte_result_t astarte_pairing_get_broker_url(
  * @param[in] crt_pem_size Size of the output buffer for the PEM certificate.
  * @return ASTARTE_RESULT_OK if successful, otherwise an error code.
  */
-astarte_result_t astarte_pairing_get_client_certificate(int32_t timeout_ms, const char *cred_secr,
-    unsigned char *privkey_pem, size_t privkey_pem_size, char *crt_pem, size_t crt_pem_size);
+astarte_result_t astarte_pairing_get_client_certificate(int32_t timeout_ms, const char *device_id,
+    const char *cred_secr, unsigned char *privkey_pem, size_t privkey_pem_size, char *crt_pem,
+    size_t crt_pem_size);
 
 /**
  * @brief Fetch the client x509 certificate from Astarte.
@@ -66,12 +69,13 @@ astarte_result_t astarte_pairing_get_client_certificate(int32_t timeout_ms, cons
  * @warning This is often a very memory intensive operation, more than 20kB of stack are required.
  *
  * @param[in] timeout_ms Timeout to use for the HTTP operations in ms.
+ * @param[in] device_id Unique identifier to use to register the device instance.
  * @param[in] cred_secr Credential secret to use as authorization token.
  * @param[in] crt_pem Input buffer containing the PEM certificate to verify.
  * @return ASTARTE_RESULT_OK if successful, otherwise an error code.
  */
 astarte_result_t astarte_pairing_verify_client_certificate(
-    int32_t timeout_ms, const char *cred_secr, const char *crt_pem);
+    int32_t timeout_ms, const char *device_id, const char *cred_secr, const char *crt_pem);
 
 #ifdef __cplusplus
 }